+++ cli/cli/App.cs
@@ -4,6 +4,7 @@ using Beamable.Common.Api;
 using Beamable.Common.Api.Auth;
 using Beamable.Common.Api.Realms;
 using Beamable.Common.Dependencies;
+using Beamable.Server;
 using cli.Commands.Project;
 using cli.Content;
 using cli.Dotnet;
@@ -20,6 +21,7 @@ using System.CommandLine;
 using System.CommandLine.Builder;
 using System.CommandLine.Invocation;
 using System.CommandLine.Parsing;
+using UnityEngine;
 
 namespace cli;
 
@@ -52,6 +54,7 @@ public class App
 
 		BeamableLogProvider.Provider = new CliSerilogProvider();
 		CliSerilogProvider.LogContext.Value = Log.Logger;
+		Debug.Instance = new BeamableLoggerDebug();
 	}
 
 	/// <summary>