+++ cli/cli/Commands/Profiling/RunNBomberCommand.cs
@@ -15,6 +15,9 @@ public class RunNBomberCommandArgs : CommandArgs
 
 	public bool includePrefix;
 	public int rps;
+	public int duration;
+	public string authHeader;
+	public bool includeAuth;
 }
 
 public class RunNBomberCommand : AppCommand<RunNBomberCommandArgs>
@@ -29,7 +32,10 @@ public class RunNBomberCommand : AppCommand<RunNBomberCommandArgs>
 		AddArgument(new Argument<string>("method", "The method name in the service to stress test"), (args, i) => args.method = i);
 		AddArgument(new Argument<string>("body", "The json body for each request"), (args, i) => args.jsonBody = i);
 		AddOption(new Option<bool>("--include-prefix", () => true, "If true, the generated .env file will include the local machine name as prefix"), (args, i) => args.includePrefix = i);
+		AddOption(new Option<bool>("--include-auth", () => true, "If true, the requests will be given the CLI's auth token"), (args, i) => args.includePrefix = i);
 		AddOption(new Option<int>("--rps", () => 50, "The requested requests per second for the test"), (args, i) => args.rps = i);
+		AddOption(new Option<int>("--duration", () => 30, "How long to run the test for"), (args, i) => args.duration = i);
+		AddOption(new Option<string>("--auth", "Include an auth header. This will override the --include-auth flag"), (args, i) => args.authHeader = i);
 	}
 
 	public override Task Handle(RunNBomberCommandArgs args)
@@ -50,6 +56,15 @@ public class RunNBomberCommand : AppCommand<RunNBomberCommandArgs>
 						.WithHeader("X-DE-SCOPE", scope)
 						.WithBody(new StringContent(args.jsonBody));
 
+				if (args.includeAuth)
+				{
+					request = request.WithHeader("Authorization", $"Bearer {args.AppContext.Token.Token}");
+				}
+				if (!string.IsNullOrEmpty(args.authHeader))
+				{
+					request = request.WithHeader("Authorization", "Bearer 5d938ebe-9cbd-41d6-b077-d8dd89b99819");
+				}
+				
 				// HttpCompletionOption: https://learn.microsoft.com/en-us/dotnet/api/system.net.http.httpcompletionoption?view=net-7.0
 
 				var clientArgs = new HttpClientArgs(
@@ -65,7 +80,7 @@ public class RunNBomberCommand : AppCommand<RunNBomberCommandArgs>
 			.WithMaxFailCount(1)
 
 			.WithLoadSimulations(Simulation.Inject(rate: args.rps, interval: TimeSpan.FromSeconds(1),
-				during: TimeSpan.FromSeconds(30)))
+				during: TimeSpan.FromSeconds(args.duration)))
 		;
 
 		NBomberRunner